[Chore] #227 placeId Long 타입으로 변경

+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/data/dto/remote/response/searchplace/list/PlaceRes.kt
@@ -10,5 +10,5 @@ data class PlaceRes(
     val mapX: String,
     val mapY: String,
     val name: String,
-    val placeId: Int
+    val placeId: Long
 )
