+++ DaOnGil/app/src/main/java/kr/tekit/lion/daongil/domain/model/RecommendPlace.kt
@@ -5,5 +5,5 @@ data class RecommendPlace (
     val disability: List<String>,
     val image: String,
     val name: String,
-    val placeId: Int
+    val placeId: Long
 )
